code reformatting + mqtt basic functionality
