+++ firmware/src/SW_Infared.cpp
@@ -9,7 +9,7 @@
 */
 
 #include "../features.h"
-#include "../src/pins.h"
+#include "../src/def_pins.h"
 #include "../src/SW_Infared.h"
 
 //------------------------------------------------------------------------------
