+++ firmware/src/SW_Infared.h
@@ -20,7 +20,7 @@
 #include "esp32-hal.h"
 #endif
 
-#include "../src/pins.h"
+#include "../src/def_pins.h"
 
 #define IR_SENSOR_COUNT 4
 #define IR_BIT_LEN 32
