+++ lib/PubSubClient/src/PubSubClient.cpp
@@ -453,7 +453,7 @@ boolean PubSubClient::beginPublish(const char* topic, unsigned int plength, bool
         // Send the header and variable length field
         uint16_t length = MQTT_MAX_HEADER_SIZE;
         length = writeString(topic,buffer,length);
-        uint16_t i;
+        //uint16_t i;  // Commented by Nuwan; due to no use
         uint8_t header = MQTTPUBLISH;
         if (retained) {
             header |= 1;
