+++ platformio.ini
@@ -8,6 +8,10 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
+[platformio]
+src_dir = firmware
+
+
 [env:nodemcu-32s]
 platform = espressif32
 board = nodemcu-32s
